Rename input -> prompt, output -> completion (#9)
Input and Output are generic names that sometimes lead to confusion. Usually, when we say "inputs" we are specifically talking about prompts to LLMs and GenAI-based products. And usually, when we say "outputs" we are specifically talking about the completions that these endpoints provide.

Hence the rename: input -> prompt, output -> completion

In particular we changed:

API routes
SQLalchemy dbs
Data Model schema
React Frontend
...to reflect this change.